backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Load .env file

class Settings:
    NEO4J_URI = os.getenv("NEO4J_URI")
    NEO4J_USER = os.getenv("NEO4J_USER")
    NEO4J_PASS = os.getenv("NEO4J_PASS")
    NEO4J_DATABASE = os.getenv("NEO4J_DATABASE")
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
    TWILIO_ACCOUNT_SID = os.getenv("TWILIO_ACCOUNT_SID")
    TWILIO_AUTH_TOKEN = os.getenv("TWILIO_AUTH_TOKEN")
    TWILIO_PHONE_NUMBER = os.getenv("TWILIO_PHONE_NUMBER")
    GEOCODE_USER_AGENT = os.getenv("GEOCODE_USER_AGENT")

settings = Settings()
