requirements.txt
fastapi
uvicorn
neo4j
requests
python-dotenv
streamlit
pyvis
plotly
scikit-learn
joblib
twilio
openai
ortools
